requirements.txt
torch>=2.0.0
torchvision
transformers
pyyaml
pillow